Initial commit on this long living branch.
Due to compatibility issues in the usage of pyinstaller and scrpts importing matplotlib, this branch was created. Calls of matplotlib and therefore the module plottery were removed. Until further updates on the designated packages, this branch shall be used for creating an exe with pyinstaller. This makes this branch the distribution branch in some way.